pyproject.toml
[project]
name = "discourse-kg"
version = "0.1.0"
dependencies = [
    "python-dotenv>=1.0.0",
    "neo4j>=5.17.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "nltk>=3.8.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0.0",
    "prefect>=3.0.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "pyprojroot>=0.2.0",
    "tqdm>=4.64.0",
    "pandas>=1.5.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]

[tool.setuptools]
packages = { find = {} }

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
